requirements.txt
xmltodict
PyYAML
msgpack
matplotlib